kernel-hal: refactor for riscv bare-metal
