+++ kernel-hal/Cargo.toml
@@ -9,6 +9,8 @@ description = "Kernel HAL interface definations."
 
 [features]
 libos = ["async-std"]
+board_qemu = []
+board_d1 = []
 
 [dependencies]
 log = "0.4"
@@ -50,5 +52,4 @@ x86-smpboot = { git = "https://github.com/rcore-os/x86-smpboot", rev = "43ffedf"
 
 # Bare-metal mode on riscv64
 [target.'cfg(all(target_os = "none", target_arch = "riscv64"))'.dependencies]
-# riscv = { git = "https://github.com/rcore-os/riscv", features = ["inline-asm"], rev = "0074cbc" }
-# volatile = "0.2"
+riscv = { git = "https://github.com/rcore-os/riscv", features = ["inline-asm"], rev = "0074cbc" }
