+++ kernel-hal/src/bare/arch/riscv/consts.rs
@@ -23,32 +23,31 @@ pub const MMIO_MTIMECMP0: *mut u64 = 0x0200_4000usize as *mut u64;
 pub const MMIO_MTIME: *const u64 = 0x0200_BFF8 as *const u64;
 
 #[cfg(feature = "board_qemu")]
-pub const UART_BASE:       usize = 0x10000000;
+pub const UART_BASE: usize = 0x10000000;
 #[cfg(feature = "board_qemu")]
-pub const UART0_INT_NUM:     u32 = 10;
+pub const UART0_INT_NUM: u32 = 10;
 #[cfg(feature = "board_qemu")]
-pub const PLIC_PRIORITY:   usize = 0x0c000000;
+pub const PLIC_PRIORITY: usize = 0x0c000000;
 #[cfg(feature = "board_qemu")]
-pub const PLIC_PENDING:    usize = 0x0c001000;
+pub const PLIC_PENDING: usize = 0x0c001000;
 #[cfg(feature = "board_qemu")]
 pub const PLIC_INT_ENABLE: usize = 0x0c002080;
 #[cfg(feature = "board_qemu")]
-pub const PLIC_THRESHOLD:  usize = 0x0c201000;
+pub const PLIC_THRESHOLD: usize = 0x0c201000;
 #[cfg(feature = "board_qemu")]
-pub const PLIC_CLAIM:      usize = 0x0c201004;
+pub const PLIC_CLAIM: usize = 0x0c201004;
 
 #[cfg(feature = "board_d1")]
-pub const UART_BASE:       usize = 0x02500000;
+pub const UART_BASE: usize = 0x02500000;
 #[cfg(feature = "board_d1")]
-pub const UART0_INT_NUM:     u32 = 18;
+pub const UART0_INT_NUM: u32 = 18;
 #[cfg(feature = "board_d1")]
-pub const PLIC_PRIORITY:   usize = 0x1000_0000;
+pub const PLIC_PRIORITY: usize = 0x1000_0000;
 #[cfg(feature = "board_d1")]
-pub const PLIC_PENDING:    usize = 0x1000_1000;
+pub const PLIC_PENDING: usize = 0x1000_1000;
 #[cfg(feature = "board_d1")]
 pub const PLIC_INT_ENABLE: usize = 0x1000_2080;
 #[cfg(feature = "board_d1")]
-pub const PLIC_THRESHOLD:  usize = 0x1020_1000;
+pub const PLIC_THRESHOLD: usize = 0x1020_1000;
 #[cfg(feature = "board_d1")]
-pub const PLIC_CLAIM:      usize = 0x1020_1004;
-
+pub const PLIC_CLAIM: usize = 0x1020_1004;
