+++ kernel-hal/src/bare/arch/riscv/cpu.rs
@@ -1,0 +1,8 @@
+hal_fn_impl! {
+    impl mod crate::defs::cpu {
+        fn cpu_frequency() -> u16 {
+            const DEFAULT: u16 = 2600;
+            DEFAULT
+        }
+    }
+}
