+++ kernel-hal/src/bare/arch/riscv/mem.rs
@@ -1,0 +1,3 @@
+pub fn frame_flush(_target: crate::PhysAddr) {
+    unimplemented!()
+}
