+++ zCore/Cargo.toml
@@ -8,8 +8,8 @@ edition = "2018"
 
 [features]
 graphic = []
-board_qemu = [] # "kernel-hal-bare/board_qemu"]
-board_d1 = ["link_user_img"] #, "kernel-hal-bare/board_d1"]
+board_qemu = ["link_user_img", "kernel-hal/board_qemu"]
+board_d1 = ["link_user_img", "kernel-hal/board_d1"]
 link_user_img = []
 zircon = ["zircon-loader"]
 linux = ["linux-loader", "linux-object", "rcore-fs-sfs"]
