+++ zCore/src/arch/riscv/boot/kernel-vars.ld
@@ -1,0 +1,2 @@
+/* Generated by build.rs. DO NOT EDIT. */
+PROVIDE_HIDDEN(BASE_ADDRESS = 0xffffffff80200000);
