+++ zircon-syscall/src/debug.rs
@@ -6,7 +6,7 @@ impl Syscall<'_> {
     pub fn sys_debug_write(&self, buf: UserInPtr<u8>, len: usize) -> ZxResult {
         info!("debug.write: buf=({:?}; {:#x})", buf, len);
         let data = buf.read_array(len)?;
-        kernel_hal::serial::print_str(core::str::from_utf8(&data).unwrap());
+        kernel_hal::serial::serial_write(core::str::from_utf8(&data).unwrap());
         Ok(())
     }
