+++ zircon-syscall/src/debuglog.rs
@@ -54,9 +54,9 @@ impl Syscall<'_> {
         let dlog = proc.get_object_with_rights::<DebugLog>(handle_value, Rights::WRITE)?;
         dlog.write(Severity::Info, options, self.thread.id(), proc.id(), &data);
         // print to kernel console
-        kernel_hal::serial::print_str(&data);
+        kernel_hal::serial::serial_write(&data);
         if data.as_bytes().last() != Some(&b'\n') {
-            kernel_hal::serial::print_str("\n");
+            kernel_hal::serial::serial_write("\n");
         }
         Ok(())
     }
